rag_app/requirements.txt
streamlit
sentence-transformers
groq
python-dotenv
endee
pypdf